sample structure made
