--- src/Model/Buyer.java
@@ -1,10 +1,0 @@
-package Model;
-
-public class Buyer extends User{
-
-    public Buyer(String name, String address, String cardnum, String clientID) {
-        super(name, address, cardnum, clientID);
-    }
-
-
-}
