--- src/Model/Seller.java
@@ -1,15 +1,0 @@
-package Model;
-
-import java.util.List;
-
-public class Seller extends User{
-
-
-    public Seller(String name, String address, String cardnum, String clientID) {
-        super(name, address, cardnum, clientID);
-
-    }
-
-
-
-}
