--- src/Service/AuctionService.java
@@ -1,8 +1,0 @@
-package Service;
-
-public class AuctionService {
-
-    //create auction
-    //update auction
-
-}
