+++ src/main/java/com/Model/Buyer.java
@@ -1,0 +1,10 @@
+package com.Model;
+
+public class Buyer extends User{
+
+    public Buyer(String name, String address, String cardnum, String clientID) {
+        super(name, address, cardnum, clientID);
+    }
+
+
+}
