+++ src/main/java/com/Model/Seller.java
@@ -1,0 +1,15 @@
+package com.Model;
+
+import java.util.List;
+
+public class Seller extends User{
+
+
+    public Seller(String name, String address, String cardnum, String clientID) {
+        super(name, address, cardnum, clientID);
+
+    }
+
+
+
+}
