+++ src/main/java/com/Service/AuctionService.java
@@ -1,0 +1,8 @@
+package com.Service;
+
+public class AuctionService {
+
+    //create auction
+    //update auction
+
+}
