+++ src/main/java/com/Service/BidService.java
@@ -1,0 +1,22 @@
+package com.Service;
+
+import com.Model.Bid;
+import com.Repository.BidRepository;
+
+public class BidService {
+
+    //add the bid
+    // update the winning bid
+
+    public void createBid(Bid bd, BidRepository bidRepo)
+    {
+        bidRepo.getBids().add(bd);
+    }
+
+
+    public Bid getWinningBid(BidRepository bidRepo)
+    {
+        return bidRepo.getWinningBid();
+    }
+
+}
