+++ src/test/java/com/carBooking/App/AppApplicationTests.java
@@ -1,0 +1,13 @@
+package com.carBooking.App;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AppApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
